PlagiarismApp/requirements.txt
amqp==5.2.0
annotated-types==0.6.0
asgiref==3.8.1
billiard==4.2.0
blis==0.7.11
Brotli==1.1.0
catalogue==2.0.10
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpathlib==0.16.0
confection==0.1.4
cssselect2==0.7.0
cymem==2.0.8
datasketch==1.6.4
Django==5.0.3
django-ranged-response==0.2.0
django-simple-captcha==0.6.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
fonttools==4.50.0
gensim==4.3.2
html5lib==1.1
idna==3.6
Jinja2==3.1.3
joblib==1.3.2
kombu==5.3.6
langcodes==3.3.0
MarkupSafe==2.1.5
murmurhash==1.0.10
mysqlclient==2.2.4
nltk==3.8.1
numpy==1.26.4
packaging==24.0
pillow==10.2.0
preshed==3.0.9
prompt-toolkit==3.0.43
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
pydyf==0.9.0
pylovepdf==1.3.2
PyPDF2==3.0.1
pyphen==0.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-magic==0.4.27
redis==5.0.3
regex==2023.12.25
requests==2.31.0
scikit-learn==1.4.1.post1
scipy==1.12.0
setuptools==69.2.0
six==1.16.0
smart-open==6.4.0
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.4.4
srsly==2.4.8
thinc==8.2.3
threadpoolctl==3.4.0
tinycss2==1.2.1
tqdm==4.66.2
typer==0.9.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wasabi==1.1.2
wcwidth==0.2.13
weasel==0.3.4
weasyprint==61.2
webencodings==0.5.1
zopfli==0.2.3
